fix: explicitly set core.abbrev=8 while getting hash
The core.abbrev configuration in git allows the user to set the short
hash to a different length. This is common to do for a lot of developers
in large projects such as the Linux Kernel where a longer short hash is
desired due to the number of commits. Currently, if the user has this
configuration set, the build will fail if -Werror=overflow is set due to
the larger value given by `git rev-parse`.

Fix the issue by explicitly setting core.abbrev=8 when running
rev-parse.

Signed-off-by: Gary Kim <gary@garykim.dev>